verification_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pika==1.3.2
python-multipart==0.0.5
